Move CollaborationManagementForm into userManagement components directory
